requirement.txt
django
djangorestframework
requests
django-service-objects==0.7.0